app/requirements.txt
Flask==2.2.2
joblib==1.2.0
pandas==1.5.1
protobuf==4.21.9
python-dotenv==0.21.0
gunicorn==20.1.0
scikit_learn==1.1.3
google-cloud-storage==2.6.*
google-cloud-core==2.3.*